Fixed direnv-mode for NixOS
Direnv overrides the path where run-time executable are
stored. Therefore Emacs does not find executable if direnv is
loaded. There are two possibilities to fix this
1) Add symlinks for all executable to the Emacs libexec path (requires
recompilation)
2) Change exec-directory which allows to add symlinks without
recompilation.

The second approach requires a configuration file which contains the
new path. This also allows the set fixed paths which is required for
some modules.